Better handling of the missing offsets
If Kafka topic that this consumer operates on has a retention policy, in some situations there can be gaps is offsets we read from Kafka.
These gaps make some pages to not get completed and therefore the consumer group's offset won't get committed properly.
Eventually these not-completed pages cause the consumer to not progress and then the consumer stops consuming records and hangs forever.